Fixes for database (description in commit info)
If projection field is present and _ts_meta isn't specified then TopicStore will assume it is a new message and reconstruct now. So we force it to always retrieve this meta data to ensure messages are always reconstructed correctly.

On slow connections auto-fetching gridFS or blob data can be a dealbreaker so a "skip_fetch_binary" flag has been added.